// src/app/modules/delivery/manager/models/manager.model.ts
export class ManagerModel {

    id: string;
    establishmentId: string;
    establishmentName: string;
    zoneId: string;
    status: number;
    isAutomatic: boolean;
    itsWeekly: boolean;
    factor: number;
    driver: number;
    capacity: number;
    estimatedTime: number;
    appOrders: number;
    cfiOrders: number;
    webOrders: number;
    totalOrders: number;
}